requirements.txt
azure-functions==1.21.3
pillow==11.0.0
